KindnessCafe/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
Babel==2.8.0
CacheControl==0.12.6
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
DateTime==4.3
defusedxml==0.6.0
distlib==0.3.0
distro==1.5.0
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.41.0
django-appconf==1.0.4
django-bootstrap-modal-forms==1.5.0
django-class-based-auth-views==0.4
django-paypal==1.0.0
django-phonenumber-field==4.0.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
ipaddr==2.2.0
isort==4.3.21
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
phonenumbers==8.12.0
Pillow==7.0.0
progress==1.5
psycopg2==2.8.5
pylint==2.4.4
pyparsing==2.4.7
python3-openid==3.1.0
pytoml==0.1.21
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.0.1
wrapt==1.11.2
zope.interface==5.1.0
